(bugfix/j-crick-fe-challenge): Refactor gulp file, add linting.
