+++ package.json
@@ -23,12 +23,16 @@
     ]
   },
   "devDependencies": {
-    "babel-preset-es2015": "^6.9.0",
+    "babel-cli": "^6.11.4",
+    "babel-preset-es2015": "^6.13.2",
     "babel-preset-stage-0": "^6.5.0",
     "babelify": "^7.3.0",
+    "browser-run": "^3.2.0",
     "browser-sync": "^2.13.0",
     "browserify": "^13.0.1",
+    "cheerio": "^0.20.0",
     "del": "^2.2.1",
+    "faucet": "0.0.1",
     "gulp": "^3.9.1",
     "gulp-autoprefixer": "^3.1.0",
     "gulp-notify": "^2.2.0",
@@ -36,6 +40,7 @@
     "gulp-sass": "^2.3.2",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-uglify": "^1.5.4",
+    "gulp-xo": "^0.12.0",
     "gutil": "^1.6.4",
     "jquery": "^3.1.0",
     "path": "^0.12.7",
