--- test/test.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8"/>
-    <title>Tests</title>
-</head>
-<body>
-<!--Required for Phantom JS 1.x -->
-<script src="../node_modules/es5-shim/es5-shim.min.js"></script>
-<script src="../node_modules/steal/steal.js"
-        data-main="test/tests">
-</script>
-</body>
-</html>
