--- test/thing.js
@@ -1,16 +1,0 @@
-
-describe('First Test', function () {
-	beforeEach(function () {
-	});
-
-	afterEach(function () {
-	});
-
-	describe('The Test', function () {
-
-		it('defaults', function () {
-
-		});
-	});
-
-});
